chore: cleanup so that no magic numbers by adding DEFAULT_LIMIT and move filters to seperate file
