+++ electron/src/chatBro/constants/defaultFilters.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_LIMIT = 15;
