+++ electron/src/chatBro/queries/WordOrEmoji/WordOrEmoji.ts
@@ -1,46 +1,16 @@
-import * as _ from 'lodash';
 import * as sqlite3 from 'sqlite3';
-import { emojis } from '../../constants/emojis';
+import { DEFAULT_LIMIT } from '../../constants/defaultFilters';
 
 import * as sqlite3Wrapper from '../../../utils/initUtils/sqliteWrapper';
 import { ChatTableNames } from '../../../tables';
 import { Columns } from '../../../tables/Core/Count';
-
-function isFromMeFilter(filters: WordOrEmojiTypes.Filters): string {
-  if (filters.isFromMe === true) {
-    return `${Columns.IS_FROM_ME} = 1`;
-  }
-  return `${Columns.IS_FROM_ME} = 0`;
-}
-
-function wordFilter(filters: WordOrEmojiTypes.Filters): string | undefined {
-  if (_.isEmpty(filters.word)) {
-    return undefined;
-  }
-  return `${Columns.WORD} = "${filters.word}"`;
-}
-
-function isEmojiFilter(filters: WordOrEmojiTypes.Filters): string {
-  if (filters.isEmoji === true) {
-    return `TRIM(${Columns.WORD}) IN (${emojis})`;
-  }
-  return `TRIM(${Columns.WORD}) NOT IN (${emojis})`;
-}
-
-// Attaches each filter in a combined WHERE clause.
-function getAllFilters(filters: WordOrEmojiTypes.Filters): string {
-  const isEmoji = isEmojiFilter(filters);
-  const isFromMe = isFromMeFilter(filters);
-  const word = wordFilter(filters);
-  const filtersArray = _.compact([isFromMe, word, isEmoji]);
-  return !_.isEmpty(filtersArray) ? `WHERE ${filtersArray.join(' AND ')}` : '';
-}
+import getAllFilters from './filters';
 
 export async function queryEmojiOrWordCounts(
   db: sqlite3.Database,
   filters: WordOrEmojiTypes.Filters = { isEmoji: false, isFromMe: true }
 ): Promise<WordOrEmojiTypes.Results> {
-  const limit = filters.limit || 15;
+  const limit = filters.limit || DEFAULT_LIMIT;
   const allFilters = getAllFilters(filters);
   const query = `
     SELECT
